requirements.txt
# Telegram bot (ваша проверенная версия)
python-telegram-bot==20.7

# Database (обновим алембик для совместимости)
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Caching and queues
redis==5.0.1
celery==5.3.4
aioredis==2.0.1

# AI and ML stack (ваши совместимые версии)
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.63
langchain-openai==0.2.8
langgraph==0.2.39
langsmith==0.1.147

# RAG и векторные БД (ваши проверенные версии)
chromadb==0.5.20
langchain-chroma==0.1.4
faiss-cpu==1.11.0
sentence-transformers==2.6.1

# Data processing and validation (ИСПРАВЛЕННЫЕ ВЕРСИИ)
numpy==1.26.4
scikit-learn==1.6.1
pydantic==2.11.4
pydantic-settings>=2.4.0,<3.0.0
jsonlines==4.0.0

# Configuration and environment (ваши версии)
python-dotenv==1.0.0
pyyaml==6.0.2

# Logging and monitoring (ваша версия + дополнения)
coloredlogs==15.0.1
structlog==23.2.0
loguru==0.7.2

# HTTP and async utilities (ваши проверенные версии)
aiohttp==3.11.18
httpx==0.25.2
aiofiles==23.2.1
tenacity==8.2.3

# Progress tracking (ваша версия)
tqdm==4.67.1

# Scheduler and background tasks
apscheduler==3.10.4
croniter==2.0.1

# Text processing
nltk==3.8.1

# Development tools (ваши версии + дополнения)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Optional: FastAPI for analytics dashboard
fastapi==0.108.0
uvicorn==0.25.0
jinja2==3.1.2

# Optional: Monitoring
prometheus-client==0.19.0

# Configuration and environment (ваши версии)
python-dotenv==1.0.0
pyyaml==6.0.2

# Logging and monitoring (ваша версия + дополнения)
coloredlogs==15.0.1
structlog==23.2.0
loguru==0.7.2

# HTTP and async utilities (ваши проверенные версии)
aiohttp==3.11.18
httpx==0.25.2
aiofiles==23.2.1
tenacity==8.2.3

# Progress tracking (ваша версия)
tqdm==4.67.1

# Scheduler and background tasks
apscheduler==3.10.4
croniter==2.0.1

# Text processing
nltk==3.8.1

# Development tools (ваши версии + дополнения)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Optional: FastAPI for analytics dashboard
fastapi==0.108.0
uvicorn==0.25.0
jinja2==3.1.2

# Optional: Monitoring
prometheus-client==0.19.0